Fix data hook bug
ERROR: When apply the paradigm that inherit from BaseTimeEncodingParadigm, using the data hook may cause a error. In the last version, the data hook input X is a Dict type with a deep list as value. This X can not be filtered using Scipy filter.

SOLUTION: Considering the purpose of the  data hook is do some process on single epoch (minimal data unit), I changed its input as unit_x and unit_y, which is a 2_D array-like type that can be filtered.