Update of SA-xTRCA
I added the SA-xTRCA class in cca.py and implemented a demo for phase alignment and classification using SA-xTRCA. In addition, I provided a portion of the experimental data (from 10 subjects) used in my research, which is included in the SA-xTRCA_demo folder. All data files are in .mat format.